refactor: Split btree_index.rs into focused submodules
Split the 1,241-line btree_index.rs file into six focused modules,
each under 300 lines, improving code organization and maintainability.

## Module Structure

- **btree_index/mod.rs** (260 lines): Core BTreeIndex struct, constructors,
  metadata I/O, and node I/O wrappers
- **btree_index/bulk_load.rs** (201 lines): Efficient bottom-up tree
  construction from sorted data
- **btree_index/insert.rs** (214 lines): Key insertion with split propagation
- **btree_index/delete.rs** (153 lines): Key deletion operations
- **btree_index/query.rs** (215 lines): Lookups, range scans, multi-key queries
- **btree_index/rebalance.rs** (271 lines): Tree rebalancing after deletions

## Changes

- Removed: `crates/vibesql-storage/src/btree/node/btree_index.rs` (1,241 lines)
- Added: 6 new module files (1,314 total lines with headers/docs)
- All 147 storage tests pass with no regressions
- Public API completely preserved
- No performance impact (pure refactoring)

## Benefits

- Each module is under 500 lines (acceptance criteria met)
- Clear separation of concerns by operation type
- Easier to navigate and maintain
- Better compile-time incremental compilation
- Reduced cognitive load when working with specific operations

Closes #1606
